+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Unreleased
 
+# 0.2.0 (2021-12-02)
+- Update `raw-window-handle` dep to `0.4.2`. Rename `macos` -> `appkit` and `ios` -> `uikit` following `raw-window-handle`.
+
 # 0.1.2
-- Update `cocoa` and `core-graphics` dependencies
+- Update `cocoa` and `core-graphics` dependencies.
 
 # 0.1.1
-- iOS: support layer from ui window
+- iOS: support layer from ui window.
 
 # 0.1.0
 - Initial release! 🎉